Reorganise controller class hierarchy
