+++ minegauler/core/split_cell/controller.py
@@ -2,21 +2,47 @@
 
 __all__ = ("GameController",)
 
-from ...shared.types import GameMode
-from ..controller import ControllerBase
+from ...shared.types import Coord_T, Difficulty, GameMode, PathLike
+from ..controller import GameControllerBase
 from .board import Board
-from .game import Game
+from .game import Game, difficulty_to_values
 from .minefield import Minefield
 from .types import Coord
 
 
-class GameController(ControllerBase):
+class _ControllerMixin:
+    """Mixin for the controller classes."""
+
+    mode = GameMode.REGULAR
+    minefield_cls = Minefield
+    board_cls = Board
+    game_cls = Game
+
+    def set_difficulty(self, difficulty: Difficulty) -> None:
+        x, y, m = difficulty_to_values(difficulty)
+        self.resize_board(x, y, m)
+
+
+class GameController(_ControllerMixin, GameControllerBase):
     """GameController for a split-cells minesweeper game."""
 
     mode = GameMode.SPLIT_CELL
     minefield_cls = Minefield
     board_cls = Board
     game_cls = Game
+
+    # --------------------------------------------------------------------------
+    # Methods triggered by user interaction
+    # --------------------------------------------------------------------------
+    def new_game(self) -> None:
+        super().new_game()
+
+    def restart_game(self) -> None:
+        """See AbstractController."""
+        super().restart_game()
+
+    def select_cell(self, coord: Coord_T) -> None:
+        pass
 
     def split_cell(self, coord: Coord) -> None:
         if coord.is_split:
@@ -32,3 +58,18 @@ class GameController(ControllerBase):
         if not coord.is_split:
             return
         super().remove_cell_flags(coord)
+
+    def chord_on_cell(self, coord: Coord_T) -> None:
+        pass
+
+    def resize_board(self, x_size: int, y_size: int, mines: int) -> None:
+        pass
+
+    def set_first_success(self, value: bool) -> None:
+        pass
+
+    def set_per_cell(self, value: int) -> None:
+        pass
+
+    def load_minefield(self, file: PathLike) -> None:
+        pass
